all little refactor : no more itinerary screen
